requirements.txt
Flask==2.2.3
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate